--- packages/pds/src/db/tables/post-embed-image.ts
@@ -1,12 +1,0 @@
-export const tableName = 'post_embed_image'
-
-export interface PostEmbedImage {
-  postUri: string
-  position: number
-  imageCid: string
-  alt: string
-}
-
-export type PartialDB = {
-  [tableName]: PostEmbedImage
-}
